Move tag data sources into domain package
