+++ internal/provider/provider.go
@@ -14,6 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/client"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/currentuser"
+	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/tag"
 )
 
 // Ensure ScaffoldingProvider satisfies various provider interfaces.
@@ -118,8 +119,8 @@ func (p *UptimeRobotProvider) DataSources(ctx context.Context) []func() datasour
 		currentuser.NewDataSource,
 		NewIPRangesDataSource,
 		NewIntegrationDataSource,
-		NewTagDataSource,
-		NewTagsDataSource,
+		tag.NewDataSource,
+		tag.NewListDataSource,
 	}
 }
