+++ internal/provider/tag/data_source_test.go
@@ -1,4 +1,4 @@
-package provider
+package tag
 
 import (
 	"strings"
